added nix-expressions
This change adds three files:
  - `create-nix-expr.sh`
  - `super-user-spark.nix`
  - `shell.nix`

`create-nix-expr.sh`: this calls `cabal2nix` (which is assumed to be in
the PATH) and generates `super-user-spark.nix`.

`shell.nix`: allows one to call `nix-shell` to be dropped into an
environment with a newly built `spark` binary. Tests will be built and
run in the process.